[IMP] Adjust logic when export image
