+++ product_export_kanban_image/wizard/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import export_product_image
+from . import export_product_image_wizard
